fixed the text scrolling bug :)
also fixed an alignment on the cursor; it looked fine when it was
after the last character, but in other spots it seemed to overlap
the previous letter. I assume this isn't desired, and it was just
a small tweak.
Please enter the commit message for your changes. Lines starting